+++ tests/testthat/test-cor_test.R
@@ -76,3 +76,10 @@ test_that("cor_test shepherd", {
   out <- cor_test(iris, "Petal.Length", "Petal.Width", method = "shepherd")
   testthat::expect_equal(out$r, as.numeric(0.94762), tol = 0.01)
 })
+
+
+test_that("cor_test blomqvist", {
+  set.seed(333)
+  out <- cor_test(iris, "Petal.Length", "Petal.Width", method = "blomqvist")
+  testthat::expect_equal(out$r, as.numeric(0.9066667), tol = 0.01)
+})